Added JSR/JSRR - someone please double check the specs and make sure I am understanding the zero extension they are talking about?

+++ src/summixSimulator/SummiX_Machine.java
@@ -106,7 +106,7 @@ public class SummiX_Machine {
 		return this.pc;
 	}
 	
-	public void setRegister(short register, short data) {
+	public void setRegister(int register, short data) {
 		/**
 		 * Sets the given register with given data and updates the
 		 * CCR accordingly
